Refactor import paths in blockController, blockchainController, and transactionsController
